Move node main package
