+++ node/bad_peer_test.go
